use ambient declarations instead of having to change the import
