+++ templates/index.ts
@@ -5,31 +5,6 @@
 import { TemplateFunction } from "../src/lib/createAdapter";
 
 const templates: { name: string, templateFunction: TemplateFunction }[] = [
-	{ name: "LICENSE.ts", templateFunction: require("./LICENSE") },
-	{ name: "README.md.ts", templateFunction: require("./README.md") },
-	{ name: "_babelrc.ts", templateFunction: require("./_babelrc") },
-	{ name: "_create-adapter.json.ts", templateFunction: require("./_create-adapter.json") },
-	{ name: "_devcontainer/README.md.ts", templateFunction: require("./_devcontainer/README.md") },
-	{ name: "_devcontainer/devcontainer.json.ts", templateFunction: require("./_devcontainer/devcontainer.json") },
-	{ name: "_devcontainer/docker-compose.yml.ts", templateFunction: require("./_devcontainer/docker-compose.yml") },
-	{ name: "_devcontainer/iobroker/_Dockerfile.ts", templateFunction: require("./_devcontainer/iobroker/_Dockerfile") },
-	{ name: "_devcontainer/nginx/nginx.conf.ts", templateFunction: require("./_devcontainer/nginx/nginx.conf") },
-	{ name: "_devcontainer/parcel/_Dockerfile.ts", templateFunction: require("./_devcontainer/parcel/_Dockerfile") },
-	{ name: "_devcontainer/parcel/run.sh.ts", templateFunction: require("./_devcontainer/parcel/run.sh") },
-	{ name: "_eslintrc_javascript.json.ts", templateFunction: require("./_eslintrc_javascript.json") },
-	{ name: "_eslintrc_typescript.js.ts", templateFunction: require("./_eslintrc_typescript.js") },
-	{ name: "_github/ISSUE_TEMPLATE/bug_report.md.ts", templateFunction: require("./_github/ISSUE_TEMPLATE/bug_report.md") },
-	{ name: "_github/auto-merge.yml.ts", templateFunction: require("./_github/auto-merge.yml") },
-	{ name: "_github/dependabot.yml.ts", templateFunction: require("./_github/dependabot.yml") },
-	{ name: "_github/workflows/dependabot-auto-merge.yml.ts", templateFunction: require("./_github/workflows/dependabot-auto-merge.yml") },
-	{ name: "_github/workflows/test-and-release.yml.ts", templateFunction: require("./_github/workflows/test-and-release.yml") },
-	{ name: "_gitignore.ts", templateFunction: require("./_gitignore") },
-	{ name: "_npmignore.ts", templateFunction: require("./_npmignore") },
-	{ name: "_prettierignore.ts", templateFunction: require("./_prettierignore") },
-	{ name: "_prettierrc.js.ts", templateFunction: require("./_prettierrc.js") },
-	{ name: "_travis.yml.ts", templateFunction: require("./_travis.yml") },
-	{ name: "_vscode/extensions.json.ts", templateFunction: require("./_vscode/extensions.json") },
-	{ name: "_vscode/settings.json.ts", templateFunction: require("./_vscode/settings.json") },
 	{ name: "admin/admin.d.ts.ts", templateFunction: require("./admin/admin.d.ts") },
 	{ name: "admin/custom_m.html.ts", templateFunction: require("./admin/custom_m.html") },
 	{ name: "admin/icon.png.ts", templateFunction: require("./admin/icon.png") },
@@ -40,13 +15,13 @@ const templates: { name: string, templateFunction: TemplateFunction }[] = [
 	{ name: "admin/src/i18n/en.json.ts", templateFunction: require("./admin/src/i18n/en.json") },
 	{ name: "admin/src/i18n/es.json.ts", templateFunction: require("./admin/src/i18n/es.json") },
 	{ name: "admin/src/i18n/fr.json.ts", templateFunction: require("./admin/src/i18n/fr.json") },
+	{ name: "admin/src/i18n/i18n.d.ts.ts", templateFunction: require("./admin/src/i18n/i18n.d.ts") },
 	{ name: "admin/src/i18n/it.json.ts", templateFunction: require("./admin/src/i18n/it.json") },
 	{ name: "admin/src/i18n/nl.json.ts", templateFunction: require("./admin/src/i18n/nl.json") },
 	{ name: "admin/src/i18n/pl.json.ts", templateFunction: require("./admin/src/i18n/pl.json") },
 	{ name: "admin/src/i18n/pt.json.ts", templateFunction: require("./admin/src/i18n/pt.json") },
 	{ name: "admin/src/i18n/ru.json.ts", templateFunction: require("./admin/src/i18n/ru.json") },
 	{ name: "admin/src/i18n/zh-cn.json.ts", templateFunction: require("./admin/src/i18n/zh-cn.json") },
-	{ name: "admin/src/i18n.ts.ts", templateFunction: require("./admin/src/i18n.ts") },
 	{ name: "admin/src/index.tsx_jsx.ts", templateFunction: require("./admin/src/index.tsx_jsx") },
 	{ name: "admin/style.css.ts", templateFunction: require("./admin/style.css") },
 	{ name: "admin/tab_m.html.ts", templateFunction: require("./admin/tab_m.html") },
@@ -56,10 +31,12 @@ const templates: { name: string, templateFunction: TemplateFunction }[] = [
 	{ name: "io-package.json.ts", templateFunction: require("./io-package.json") },
 	{ name: "lib/adapter-config.d.ts.ts", templateFunction: require("./lib/adapter-config.d.ts") },
 	{ name: "lib/tools.js.ts", templateFunction: require("./lib/tools.js") },
+	{ name: "LICENSE.ts", templateFunction: require("./LICENSE") },
 	{ name: "main.es6.js.ts", templateFunction: require("./main.es6.js") },
 	{ name: "main.js.ts", templateFunction: require("./main.js") },
 	{ name: "main.test.js.ts", templateFunction: require("./main.test.js") },
 	{ name: "package.json.ts", templateFunction: require("./package.json") },
+	{ name: "README.md.ts", templateFunction: require("./README.md") },
 	{ name: "src/lib/tools.ts.ts", templateFunction: require("./src/lib/tools.ts") },
 	{ name: "src/main.es6.ts", templateFunction: require("./src/main.es6") },
 	{ name: "src/main.test.ts.ts", templateFunction: require("./src/main.test.ts") },
@@ -76,5 +53,28 @@ const templates: { name: string, templateFunction: TemplateFunction }[] = [
 	{ name: "widgets/style.css.ts", templateFunction: require("./widgets/style.css") },
 	{ name: "widgets/template.html.ts", templateFunction: require("./widgets/template.html") },
 	{ name: "widgets/template.js.ts", templateFunction: require("./widgets/template.js") },
+	{ name: "_babelrc.ts", templateFunction: require("./_babelrc") },
+	{ name: "_create-adapter.json.ts", templateFunction: require("./_create-adapter.json") },
+	{ name: "_devcontainer/devcontainer.json.ts", templateFunction: require("./_devcontainer/devcontainer.json") },
+	{ name: "_devcontainer/docker-compose.yml.ts", templateFunction: require("./_devcontainer/docker-compose.yml") },
+	{ name: "_devcontainer/iobroker/_Dockerfile.ts", templateFunction: require("./_devcontainer/iobroker/_Dockerfile") },
+	{ name: "_devcontainer/nginx/nginx.conf.ts", templateFunction: require("./_devcontainer/nginx/nginx.conf") },
+	{ name: "_devcontainer/parcel/run.sh.ts", templateFunction: require("./_devcontainer/parcel/run.sh") },
+	{ name: "_devcontainer/parcel/_Dockerfile.ts", templateFunction: require("./_devcontainer/parcel/_Dockerfile") },
+	{ name: "_devcontainer/README.md.ts", templateFunction: require("./_devcontainer/README.md") },
+	{ name: "_eslintrc_javascript.json.ts", templateFunction: require("./_eslintrc_javascript.json") },
+	{ name: "_eslintrc_typescript.js.ts", templateFunction: require("./_eslintrc_typescript.js") },
+	{ name: "_github/auto-merge.yml.ts", templateFunction: require("./_github/auto-merge.yml") },
+	{ name: "_github/dependabot.yml.ts", templateFunction: require("./_github/dependabot.yml") },
+	{ name: "_github/ISSUE_TEMPLATE/bug_report.md.ts", templateFunction: require("./_github/ISSUE_TEMPLATE/bug_report.md") },
+	{ name: "_github/workflows/dependabot-auto-merge.yml.ts", templateFunction: require("./_github/workflows/dependabot-auto-merge.yml") },
+	{ name: "_github/workflows/test-and-release.yml.ts", templateFunction: require("./_github/workflows/test-and-release.yml") },
+	{ name: "_gitignore.ts", templateFunction: require("./_gitignore") },
+	{ name: "_npmignore.ts", templateFunction: require("./_npmignore") },
+	{ name: "_prettierignore.ts", templateFunction: require("./_prettierignore") },
+	{ name: "_prettierrc.js.ts", templateFunction: require("./_prettierrc.js") },
+	{ name: "_travis.yml.ts", templateFunction: require("./_travis.yml") },
+	{ name: "_vscode/extensions.json.ts", templateFunction: require("./_vscode/extensions.json") },
+	{ name: "_vscode/settings.json.ts", templateFunction: require("./_vscode/settings.json") },
 ];
 export = templates;
